requirements.txt
numpy==1.15.2
scipy==1.1.0
nltk==3.3
treetaggerwrapper==2.2.7
matplotlib==3.0.1
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Theano==1.0.4
gensim==3.7.0
tensorflow==1.12.0
